Fix potential crash in case auth token request fails.
Previously, the NSError generated by beginAuthWithCallbackURL:... if auth token or secret are missing sets the NSURLResponse object as NSLocalizedDescriptionKey. According to the docs, the object associated with NSLocalizedDescriptionKey must be a string. In particular, if you called -[NSError localizedDescription] on the object, a "unrecongnized selector 'length' sent to instance NSError" error occurred.

This commit sets a string containing a generic error message including HTTP status code as NSLocalizedDescriptionKey.